requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.0
discord.py @ git+https://github.com/Rapptz/discord.py@11278d55e80fb7b1e928ce95b2eff32a1b4e0596
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
mypy==0.971
mypy-extensions==0.4.3
python-dotenv==0.20.0
tomli==2.0.1
typing_extensions==4.3.0
yarl==1.8.1
